fix: 実行画面スピナー改善・デッドコード削除・不要スキル削除
- summaryTextにスピナーを統合し、完了時にDone表示に切り替え
- startSpinnerヘルパーに抽出（複数回呼び出しガード付き）
- loading-spinner.ts を削除（デッドコード）
- find-refactoringスキルを削除（ローカルLLMでは実用的でないため）
